Updated AuthContext and added a Routing parameter attribute
