+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Attributes/RequireQueryParameterAttribute.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.AspNetCore.Mvc.Abstractions;
+using Microsoft.AspNetCore.Mvc.ActionConstraints;
+using Microsoft.AspNetCore.Routing;
+
+namespace SFA.DAS.ProviderCommitments.Web.Attributes
+{
+    public class RequireQueryParameterAttribute : ActionMethodSelectorAttribute
+    {
+        public RequireQueryParameterAttribute(string valueName)
+        {
+            ValueName = valueName;
+        }
+        public string ValueName { get; private set; }
+        public override bool IsValidForRequest(RouteContext routeContext, ActionDescriptor action)
+        {
+            return routeContext.HttpContext.Request.Query.ContainsKey(ValueName);
+        }
+    }
+}
